requirements.txt
# requirements.txt

web3==6.11.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
eth-typing==3.5.2
eth-utils==2.3.1